Show the 'Import Wallet' and 'Create Wallet' options on the modal when activating advanced mode (BTC wallet). (#281)
* update advanced-mode-modal to have Create Wallet and Import Wallet buttons

* text added

* update advanced-mode-toggle component to enable import or create wallet whenadvanced mode turned on

* show 'Show Recovery Phrase' and 'Back Home' buttons if BTC wallet is imported before

* get previously imported wallet if recovery phrase exists

* disconnect breeze sdk if it is initialized before

+++ app/utils/breez-sdk-liquid/index.ts
@@ -31,7 +31,7 @@ import {
   ReceivePaymentResponse,
 } from "@breeztech/react-native-breez-sdk-liquid"
 
-const KEYCHAIN_MNEMONIC_KEY = "mnemonic_key"
+export const KEYCHAIN_MNEMONIC_KEY = "mnemonic_key"
 
 export let breezSDKInitialized = false
 let breezSDKInitializing: Promise<void | boolean> | null = null
@@ -94,11 +94,13 @@ const connectToSDK = async () => {
 
 export const disconnectToSDK = async () => {
   try {
-    const config = await defaultConfig(LiquidNetwork.MAINNET)
-    await disconnect()
-    await RNFS.unlink(config.workingDir)
-    breezSDKInitialized = false
-    breezSDKInitializing = null
+    if (breezSDKInitialized) {
+      const config = await defaultConfig(LiquidNetwork.MAINNET)
+      await disconnect()
+      await RNFS.unlink(config.workingDir)
+      breezSDKInitialized = false
+      breezSDKInitializing = null
+    }
   } catch (error) {
     console.error("Disconnect error: ", error)
     throw error
